Fix for 64/32 bits compilation.
git-svn-id: http://subversion.jackaudio.org/jack/jack2/trunk/jackmp@4497 0c269be4-1314-0410-8aa9-9f06e86f4224

+++ common/JackNetManager.cpp
@@ -431,7 +431,7 @@ namespace Jack
         for (int audio_port_index = 0; audio_port_index < fParams.fSendAudioChannels; audio_port_index++) {
 
         #ifdef OPTIMIZED_PROTOCOL
-            if ((long)fNetAudioCaptureBuffer->GetBuffer(audio_port_index) == -1) {
+            if ((uintptr_t)fNetAudioCaptureBuffer->GetBuffer(audio_port_index) == -1) {
                 // Port is connected on other side...
                 fNetAudioCaptureBuffer->SetBuffer(audio_port_index,
                                                 static_cast<sample_t*>(jack_port_get_buffer_nulled(fAudioCapturePorts[audio_port_index],
